# Make write faster on 64-bit platforms