@tkaitchuck 2019-04-01
Yes, that loop structure was a significant optimization in aHash. It's worth noting that this changes the value of the data being hashed. 
For example before if you hashed "abc" it would put a null at the top and feed 0abc into the hash.
With this change it will pass in "abbc" doubling the b to pad the data. This can obviously produce collisions unexpectedly. For this reason aHash hashes the length of the string in addition to the data.