@tkaitchuck 2019-04-01
Another part of what makes aHash fast is the way it reads out data. `as_array` and all those `convert()` functions would probably help here too. I can put them in their own crate if that is desirable.