@Zoxc 2019-04-01
> Yes, that loop structure was a significant optimization in aHash. It's worth noting that this changes the value of the data being hashed.

I think you mean the value of the hash changes here, which is fine for this crate.

> Another part of what makes aHash fast is the way it reads out data. `as_array` and all those `convert()` functions would probably help here too. I can put them in their own crate if that is desirable.

LLVM optimizes all the functions used here away anyway. It failed to remove the bound check when accessing the trailing bytes after the loop though, hence I used `get_unchecked` there. The same thing happens in `aHash` (for both AES and the fallback).