@tkaitchuck 2019-04-01
> It failed to remove the bound check when accessing the trailing bytes after the loop though, hence I used get_unchecked there. The same thing happens in aHash (for both AES and the fallback).

Humm, I thought I fixed that...
(How are you seeing that BTW, I found I have to disable inlining to get the generated assembly to do what I expect.)