@Zoxc 2019-04-01
@tkaitchuck I add some functions like these to inspect the generated code:
```rust
#[inline(never)]
#[no_mangle]
fn hash_test(input: &[u8]) -> u64 {
    let mut a = AHasher::new_with_keys(67, 87);
    a.write(input);
    a.finish()
}

#[inline(never)]
#[no_mangle]
fn hash_test_one(input: u64) -> u64 {
    let mut a = AHasher::new_with_keys(67, 87);
    a.write_u64(input);
    a.finish()
}
```