@Zoxc 2019-04-09
This didn't turn out to be a performance win for rustc, see https://github.com/rust-lang/rust/pull/59594.